Run automated code cleanup

+++ Codecs/Audio/Alaw/ALawEncoder.cs
@@ -45,7 +45,7 @@
                 sample = cClip;
             if (sample >= 256)
             {
-                exponent = (int)ALawCompressTable[(sample >> 8) & 0x7F];
+                exponent = ALawCompressTable[(sample >> 8) & 0x7F];
                 mantissa = (sample >> (exponent + 3)) & 0x0F;
                 compressedByte = (byte)((exponent << 4) | mantissa);
             }
